+++ apps/mobile/src/app/(home)/home/home.tsx
@@ -1,13 +1,11 @@
 import { PageLayout } from '@/components/page/page.layout';
 import { AccountsWidget } from '@/components/widgets/accounts/accounts-widget';
-import {
-  CollectiblesWidget,
-  mockCollectibles,
-  serializeCollectibles,
-} from '@/components/widgets/collectibles';
-import { TokensWidget } from '@/components/widgets/tokens';
-import { mockTotalBalance } from '@/mocks/balance.mocks';
-import { getMockTokens } from '@/mocks/tokens.mocks';
+import { AllAccountBalances, TokensWidget } from '@/components/widgets/tokens/tokens-widget';
+// import {
+//   CollectiblesWidget,
+//   mockCollectibles,
+//   serializeCollectibles,
+// } from '@/components/widgets/collectibles';
 import { useAccounts } from '@/store/accounts/accounts.read';
 import { useWallets } from '@/store/wallets/wallets.read';
 import { useLingui } from '@lingui/react';
@@ -20,11 +18,14 @@ export function Home() {
   return (
     <PageLayout>
       <AccountsWidget accounts={accounts.list} wallets={wallets.list} />
-      <TokensWidget tokens={getMockTokens()} totalBalance={mockTotalBalance} />
-      <CollectiblesWidget
+      <TokensWidget>
+        <AllAccountBalances />
+      </TokensWidget>
+
+      {/* <CollectiblesWidget
         collectibles={serializeCollectibles(mockCollectibles)}
         totalBalance={mockTotalBalance}
-      />
+      /> */}
     </PageLayout>
   );
 }