+++ apps/mobile/src/components/widgets/tokens/tokens-widget.tsx
@@ -1,38 +1,36 @@
-import React from 'react';
-
+import {
+  BitcoinBalance,
+  BitcoinBalanceByAccount,
+} from '@/features/balances/bitcoin/bitcoin-balance';
+import { TokenBalance } from '@/features/balances/token-balance';
+import { AccountId } from '@/models/domain.model';
+import { HasChildren } from '@/utils/types';
 import { t } from '@lingui/macro';
 
-import { Money } from '@leather.io/models';
-import { Flag, ItemLayout } from '@leather.io/ui/native';
-
-import { type Token } from '../../../mocks/tokens.mocks';
-import { Balance } from '../../balance/balance';
+import { getMockTokens } from '../../../mocks/tokens.mocks';
 import { Widget, WidgetHeader } from '../components/widget';
 
-interface TokensWidgetProps {
-  tokens: Token[];
-  totalBalance: Money;
-}
-
-function showChain(chain: string) {
+export function showChain(chain: string) {
   if (chain === 'Stacks blockchain' || chain === 'Bitcoin blockchain') return;
   return chain;
 }
 
-export function TokensWidget({ tokens, totalBalance }: TokensWidgetProps) {
+export function TokensWidget({ children }: HasChildren) {
   return (
     <Widget
-      header={
-        <WidgetHeader
-          title={t({
-            id: 'tokens.header_title',
-            message: 'My tokens',
-          })}
-          totalBalance={totalBalance}
-        />
-      }
+      header={<WidgetHeader title={t({ id: 'tokens.header_title', message: 'My tokens' })} />}
     >
-      {tokens.map(
+      {children}
+    </Widget>
+  );
+}
+
+export function AllAccountBalances() {
+  return (
+    <>
+      <BitcoinBalance />
+
+      {getMockTokens().map(
         ({
           availableBalance: { availableBalance },
           icon,
@@ -41,16 +39,22 @@ export function TokensWidget({ tokens, totalBalance }: TokensWidgetProps) {
           chain,
           fiatBalance,
         }) => (
-          <Flag key={ticker} img={icon} align="middle" spacing="1" reverse={false}>
-            <ItemLayout
-              titleLeft={tokenName}
-              titleRight={availableBalance && <Balance balance={availableBalance} />}
-              captionLeft={showChain(chain)}
-              captionRight={<Balance balance={fiatBalance} color="ink.text-subdued" />}
-            />
-          </Flag>
+          <TokenBalance
+            key={ticker}
+            availableBalance={availableBalance}
+            icon={icon}
+            ticker={ticker}
+            tokenName={tokenName}
+            chain={chain}
+            fiatBalance={fiatBalance}
+            showChain={showChain}
+          />
         )
       )}
-    </Widget>
+    </>
   );
+}
+
+export function AccountBalances({ fingerprint, accountIndex }: AccountId) {
+  return <BitcoinBalanceByAccount fingerprint={fingerprint} accountIndex={accountIndex} />;
 }