+++ packages/query/src/bitcoin/address/utxos-by-address.spec.tsx
@@ -4,7 +4,7 @@ import { filterUtxosWithInscriptions } from './utxos-by-address.hooks';
 
 describe(filterUtxosWithInscriptions, () => {
   test('that it filters out utxos with inscriptions so they are not spent', () => {
-    const filteredUtxos = filterUtxosWithInscriptions(mockInscriptionsList, mockUtxos);
+    const filteredUtxos = mockUtxos.filter(filterUtxosWithInscriptions(mockInscriptionsList));
     expect(filteredUtxos).toEqual([]);
   });
 });