+++ packages/utils/src/time.ts
@@ -1,4 +1,5 @@
+// WARNING: When using `setTimeout` method, there is an upper maximum
+// https://developer.mozilla.org/en-US/docs/Web/API/Window/setTimeout#maximum_delay_value
 export const fiveMinInMs = 5 * 60 * 1000;
 export const oneDayInMs = 24 * 60 * 60 * 1000;
 export const oneWeekInMs = 7 * oneDayInMs;
-export const oneMonthInMs = 30 * oneDayInMs;