+++ packages/crypto/CHANGELOG.md
@@ -72,6 +72,12 @@
   * dependencies
     * @leather.io/utils bumped to 0.16.6
 
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @leather.io/utils bumped to 0.16.7
+
 ## [1.6.0](https://github.com/leather-io/mono/compare/@leather.io/crypto-v1.5.2...@leather.io/crypto-v1.6.0) (2024-09-20)